+++ oprofile/org.eclipse.linuxtools.oprofile.ui.tests/src/org/eclipse/linuxtools/oprofile/ui/tests/TestUiDataModel.java
@@ -489,7 +489,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e1_s1_i1_s1_samples[0].toString());
 		assertNotNull(e1_s1_i1_s1_samples[0].getLabelText());
 		assertEquals(42, ((UiModelSample)e1_s1_i1_s1_samples[0]).getLine());
-		assertEquals((double)0.6341, ((UiModelSample)e1_s1_i1_s1_samples[0]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.6341, ((UiModelSample)e1_s1_i1_s1_samples[0]).getCountPercentage(), 0.0001);
 		assertNull(e1_s1_i1_s1_samples[0].getChildren());
 		assertFalse(e1_s1_i1_s1_samples[0].hasChildren());
 		assertEquals(e1_s1_i1_symbols[0], e1_s1_i1_s1_samples[0].getParent());
@@ -498,7 +498,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e1_s1_i1_s1_samples[1].toString());
 		assertNotNull(e1_s1_i1_s1_samples[1].getLabelText());
 		assertEquals(36, ((UiModelSample)e1_s1_i1_s1_samples[1]).getLine());
-		assertEquals((double)0.1951, ((UiModelSample)e1_s1_i1_s1_samples[1]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.1951, ((UiModelSample)e1_s1_i1_s1_samples[1]).getCountPercentage(), 0.0001);
 		assertNull(e1_s1_i1_s1_samples[1].getChildren());
 		assertFalse(e1_s1_i1_s1_samples[1].hasChildren());
 		assertEquals(e1_s1_i1_symbols[0], e1_s1_i1_s1_samples[1].getParent());
@@ -507,7 +507,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e1_s1_i1_s1_samples[2].toString());
 		assertNotNull(e1_s1_i1_s1_samples[2].getLabelText());
 		assertEquals(31, ((UiModelSample)e1_s1_i1_s1_samples[2]).getLine());
-		assertEquals((double)0.0488, ((UiModelSample)e1_s1_i1_s1_samples[2]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.0488, ((UiModelSample)e1_s1_i1_s1_samples[2]).getCountPercentage(), 0.0001);
 		assertNull(e1_s1_i1_s1_samples[2].getChildren());
 		assertFalse(e1_s1_i1_s1_samples[2].hasChildren());
 		assertEquals(e1_s1_i1_symbols[0], e1_s1_i1_s1_samples[2].getParent());
@@ -516,7 +516,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e1_s1_i1_s1_samples[3].toString());
 		assertNotNull(e1_s1_i1_s1_samples[3].getLabelText());
 		assertEquals(39, ((UiModelSample)e1_s1_i1_s1_samples[3]).getLine());
-		assertEquals((double)0.0, ((UiModelSample)e1_s1_i1_s1_samples[3]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.0, ((UiModelSample)e1_s1_i1_s1_samples[3]).getCountPercentage(), 0.0001);
 		assertNull(e1_s1_i1_s1_samples[3].getChildren());
 		assertFalse(e1_s1_i1_s1_samples[3].hasChildren());
 		assertEquals(e1_s1_i1_symbols[0], e1_s1_i1_s1_samples[3].getParent());
@@ -525,7 +525,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e1_s1_i1_s2_samples[0].toString());
 		assertNotNull(e1_s1_i1_s2_samples[0].getLabelText());
 		assertEquals(94, ((UiModelSample)e1_s1_i1_s2_samples[0]).getLine());
-		assertEquals((double)0.0976, ((UiModelSample)e1_s1_i1_s2_samples[0]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.0976, ((UiModelSample)e1_s1_i1_s2_samples[0]).getCountPercentage(), 0.0001);
 		assertNull(e1_s1_i1_s2_samples[0].getChildren());
 		assertFalse(e1_s1_i1_s2_samples[0].hasChildren());
 		assertEquals(e1_s1_i1_symbols[1], e1_s1_i1_s2_samples[0].getParent());
@@ -534,7 +534,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e1_s1_i1_s2_samples[1].toString());
 		assertNotNull(e1_s1_i1_s2_samples[1].getLabelText());
 		assertEquals(12, ((UiModelSample)e1_s1_i1_s2_samples[1]).getLine());
-		assertEquals((double)0.0, ((UiModelSample)e1_s1_i1_s2_samples[1]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.0, ((UiModelSample)e1_s1_i1_s2_samples[1]).getCountPercentage(), 0.0001);
 		assertNull(e1_s1_i1_s2_samples[1].getChildren());
 		assertFalse(e1_s1_i1_s2_samples[1].hasChildren());
 		assertEquals(e1_s1_i1_symbols[1], e1_s1_i1_s2_samples[1].getParent());
@@ -543,7 +543,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e1_s1_i1_s2_samples[2].toString());
 		assertNotNull(e1_s1_i1_s2_samples[2].getLabelText());
 		assertEquals(55, ((UiModelSample)e1_s1_i1_s2_samples[2]).getLine());
-		assertEquals((double)0.0, ((UiModelSample)e1_s1_i1_s2_samples[2]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.0, ((UiModelSample)e1_s1_i1_s2_samples[2]).getCountPercentage(), 0.0001);
 		assertNull(e1_s1_i1_s2_samples[2].getChildren());
 		assertFalse(e1_s1_i1_s2_samples[2].hasChildren());
 		assertEquals(e1_s1_i1_symbols[1], e1_s1_i1_s2_samples[2].getParent());
@@ -552,7 +552,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e2_s1_i1_s1_samples[0].toString());
 		assertNotNull(e2_s1_i1_s1_samples[0].getLabelText());
 		assertEquals(42, ((UiModelSample)e2_s1_i1_s1_samples[0]).getLine());
-		assertEquals((double)0.6341, ((UiModelSample)e2_s1_i1_s1_samples[0]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.6341, ((UiModelSample)e2_s1_i1_s1_samples[0]).getCountPercentage(), 0.0001);
 		assertNull(e2_s1_i1_s1_samples[0].getChildren());
 		assertFalse(e2_s1_i1_s1_samples[0].hasChildren());
 		assertEquals(e2_s1_i1_symbols[0], e2_s1_i1_s1_samples[0].getParent());
@@ -561,7 +561,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e2_s1_i1_s1_samples[1].toString());
 		assertNotNull(e2_s1_i1_s1_samples[1].getLabelText());
 		assertEquals(36, ((UiModelSample)e2_s1_i1_s1_samples[1]).getLine());
-		assertEquals((double)0.1951, ((UiModelSample)e2_s1_i1_s1_samples[1]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.1951, ((UiModelSample)e2_s1_i1_s1_samples[1]).getCountPercentage(), 0.0001);
 		assertNull(e2_s1_i1_s1_samples[1].getChildren());
 		assertFalse(e2_s1_i1_s1_samples[1].hasChildren());
 		assertEquals(e2_s1_i1_symbols[0], e2_s1_i1_s1_samples[1].getParent());
@@ -570,7 +570,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e2_s1_i1_s1_samples[2].toString());
 		assertNotNull(e2_s1_i1_s1_samples[2].getLabelText());
 		assertEquals(31, ((UiModelSample)e2_s1_i1_s1_samples[2]).getLine());
-		assertEquals((double)0.0488, ((UiModelSample)e2_s1_i1_s1_samples[2]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.0488, ((UiModelSample)e2_s1_i1_s1_samples[2]).getCountPercentage(), 0.0001);
 		assertNull(e2_s1_i1_s1_samples[2].getChildren());
 		assertFalse(e2_s1_i1_s1_samples[2].hasChildren());
 		assertEquals(e2_s1_i1_symbols[0], e2_s1_i1_s1_samples[2].getParent());
@@ -579,7 +579,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e2_s1_i1_s1_samples[3].toString());
 		assertNotNull(e2_s1_i1_s1_samples[3].getLabelText());
 		assertEquals(39, ((UiModelSample)e2_s1_i1_s1_samples[3]).getLine());
-		assertEquals((double)0.0, ((UiModelSample)e2_s1_i1_s1_samples[3]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.0, ((UiModelSample)e2_s1_i1_s1_samples[3]).getCountPercentage(), 0.0001);
 		assertNull(e2_s1_i1_s1_samples[3].getChildren());
 		assertFalse(e2_s1_i1_s1_samples[3].hasChildren());
 		assertEquals(e2_s1_i1_symbols[0], e2_s1_i1_s1_samples[3].getParent());
@@ -588,7 +588,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e2_s1_i1_s2_samples[0].toString());
 		assertNotNull(e2_s1_i1_s2_samples[0].getLabelText());
 		assertEquals(94, ((UiModelSample)e2_s1_i1_s2_samples[0]).getLine());
-		assertEquals((double)0.0976, ((UiModelSample)e2_s1_i1_s2_samples[0]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.0976, ((UiModelSample)e2_s1_i1_s2_samples[0]).getCountPercentage(), 0.0001);
 		assertNull(e2_s1_i1_s2_samples[0].getChildren());
 		assertFalse(e2_s1_i1_s2_samples[0].hasChildren());
 		assertEquals(e2_s1_i1_symbols[1], e2_s1_i1_s2_samples[0].getParent());
@@ -597,7 +597,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e2_s1_i1_s2_samples[1].toString());
 		assertNotNull(e2_s1_i1_s2_samples[1].getLabelText());
 		assertEquals(12, ((UiModelSample)e2_s1_i1_s2_samples[1]).getLine());
-		assertEquals((double)0.0, ((UiModelSample)e2_s1_i1_s2_samples[1]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.0, ((UiModelSample)e2_s1_i1_s2_samples[1]).getCountPercentage(), 0.0001);
 		assertNull(e2_s1_i1_s2_samples[1].getChildren());
 		assertFalse(e2_s1_i1_s2_samples[1].hasChildren());
 		assertEquals(e2_s1_i1_symbols[1], e2_s1_i1_s2_samples[1].getParent());
@@ -606,7 +606,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e2_s1_i1_s2_samples[2].toString());
 		assertNotNull(e2_s1_i1_s2_samples[2].getLabelText());
 		assertEquals(55, ((UiModelSample)e2_s1_i1_s2_samples[2]).getLine());
-		assertEquals((double)0.0, ((UiModelSample)e2_s1_i1_s2_samples[2]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.0, ((UiModelSample)e2_s1_i1_s2_samples[2]).getCountPercentage(), 0.0001);
 		assertNull(e2_s1_i1_s2_samples[2].getChildren());
 		assertFalse(e2_s1_i1_s2_samples[2].hasChildren());
 		assertEquals(e2_s1_i1_symbols[1], e2_s1_i1_s2_samples[2].getParent());
@@ -615,7 +615,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e2_s2_i1_s1_samples[0].toString());
 		assertNotNull(e2_s2_i1_s1_samples[0].getLabelText());
 		assertEquals(42, ((UiModelSample)e2_s2_i1_s1_samples[0]).getLine());
-		assertEquals((double)0.65, ((UiModelSample)e2_s2_i1_s1_samples[0]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.65, ((UiModelSample)e2_s2_i1_s1_samples[0]).getCountPercentage(), 0.0001);
 		assertNull(e2_s2_i1_s1_samples[0].getChildren());
 		assertFalse(e2_s2_i1_s1_samples[0].hasChildren());
 		assertEquals(e2_s2_i1_symbols[0], e2_s2_i1_s1_samples[0].getParent());
@@ -624,7 +624,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e2_s2_i1_s1_samples[1].toString());
 		assertNotNull(e2_s2_i1_s1_samples[1].getLabelText());
 		assertEquals(36, ((UiModelSample)e2_s2_i1_s1_samples[1]).getLine());
-		assertEquals((double)0.2, ((UiModelSample)e2_s2_i1_s1_samples[1]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.2, ((UiModelSample)e2_s2_i1_s1_samples[1]).getCountPercentage(), 0.0001);
 		assertNull(e2_s2_i1_s1_samples[1].getChildren());
 		assertFalse(e2_s2_i1_s1_samples[1].hasChildren());
 		assertEquals(e2_s2_i1_symbols[0], e2_s2_i1_s1_samples[1].getParent());
@@ -633,7 +633,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e2_s2_i1_s1_samples[2].toString());
 		assertNotNull(e2_s2_i1_s1_samples[2].getLabelText());
 		assertEquals(31, ((UiModelSample)e2_s2_i1_s1_samples[2]).getLine());
-		assertEquals((double)0.05, ((UiModelSample)e2_s2_i1_s1_samples[2]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.05, ((UiModelSample)e2_s2_i1_s1_samples[2]).getCountPercentage(), 0.0001);
 		assertNull(e2_s2_i1_s1_samples[2].getChildren());
 		assertFalse(e2_s2_i1_s1_samples[2].hasChildren());
 		assertEquals(e2_s2_i1_symbols[0], e2_s2_i1_s1_samples[2].getParent());
@@ -642,7 +642,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e2_s2_i1_s1_samples[3].toString());
 		assertNotNull(e2_s2_i1_s1_samples[3].getLabelText());
 		assertEquals(39, ((UiModelSample)e2_s2_i1_s1_samples[3]).getLine());
-		assertEquals((double)0.0, ((UiModelSample)e2_s2_i1_s1_samples[3]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.0, ((UiModelSample)e2_s2_i1_s1_samples[3]).getCountPercentage(), 0.0001);
 		assertNull(e2_s2_i1_s1_samples[3].getChildren());
 		assertFalse(e2_s2_i1_s1_samples[3].hasChildren());
 		assertEquals(e2_s2_i1_symbols[0], e2_s2_i1_s1_samples[3].getParent());
@@ -651,7 +651,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e2_s2_i1_s2_samples[0].toString());
 		assertNotNull(e2_s2_i1_s2_samples[0].getLabelText());
 		assertEquals(94, ((UiModelSample)e2_s2_i1_s2_samples[0]).getLine());
-		assertEquals((double)0.1, ((UiModelSample)e2_s2_i1_s2_samples[0]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.1, ((UiModelSample)e2_s2_i1_s2_samples[0]).getCountPercentage(), 0.0001);
 		assertNull(e2_s2_i1_s2_samples[0].getChildren());
 		assertFalse(e2_s2_i1_s2_samples[0].hasChildren());
 		assertEquals(e2_s2_i1_symbols[1], e2_s2_i1_s2_samples[0].getParent());
@@ -660,7 +660,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e2_s2_i1_s2_samples[1].toString());
 		assertNotNull(e2_s2_i1_s2_samples[1].getLabelText());
 		assertEquals(12, ((UiModelSample)e2_s2_i1_s2_samples[1]).getLine());
-		assertEquals((double)0.0, ((UiModelSample)e2_s2_i1_s2_samples[1]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.0, ((UiModelSample)e2_s2_i1_s2_samples[1]).getCountPercentage(), 0.0001);
 		assertNull(e2_s2_i1_s2_samples[1].getChildren());
 		assertFalse(e2_s2_i1_s2_samples[1].hasChildren());
 		assertEquals(e2_s2_i1_symbols[1], e2_s2_i1_s2_samples[1].getParent());
@@ -669,7 +669,7 @@ public class TestUiDataModel extends TestCase {
 		assertNotNull(e2_s2_i1_s2_samples[2].toString());
 		assertNotNull(e2_s2_i1_s2_samples[2].getLabelText());
 		assertEquals(55, ((UiModelSample)e2_s2_i1_s2_samples[2]).getLine());
-		assertEquals((double)0.0, ((UiModelSample)e2_s2_i1_s2_samples[2]).getCountPercentage(), (double)0.0001);
+		assertEquals(0.0, ((UiModelSample)e2_s2_i1_s2_samples[2]).getCountPercentage(), 0.0001);
 		assertNull(e2_s2_i1_s2_samples[2].getChildren());
 		assertFalse(e2_s2_i1_s2_samples[2].hasChildren());
 		assertEquals(e2_s2_i1_symbols[1], e2_s2_i1_s2_samples[2].getParent());
